fixed linter errors and removed unwanted comments
